Split calendar component into festival-calendar and resource-calendar to make different use cases more manageable
