+++ src/routes/(unauthed)/create/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-  import Calendar from '../../../components/calendar.svelte';
+  import FestivalCalendar from '../../../components/festival-calendar.svelte';
 </script>
 
 <h1>Welcome to ToolKitties! Start here to organise your own programme.</h1>
@@ -7,10 +7,8 @@
 <form>
   <input id="name" name="name" type="text" placeholder="Programme name" />
   <input id="username" name="username" type="text" placeholder="Your name" />
-  <Calendar 
-    use={"festival creation"}
-    canSelectMultiple={true} 
-    hasTimePicker={false} 
-/>
+  <FestivalCalendar 
+    canSelectMultiple={true}
+  />
   <button class="border border-black rounded p-4" type="submit">Create</button>
 </form>
